better warning for mismatched types and bad userdata casts (issue #301)

+++ src/moaicore/MOAIInputMgr.cpp
@@ -89,11 +89,11 @@ void MOAIInputMgr::EnqueuePointerEvent ( u8 deviceID, u8 sensorID, int x, int y 
 }
 
 //----------------------------------------------------------------//
-void MOAIInputMgr::EnqueueTouchEvent ( u8 deviceID, u8 sensorID, u32 touchID, bool down, float x, float y) {
+void MOAIInputMgr::EnqueueTouchEvent ( u8 deviceID, u8 sensorID, u32 touchID, bool down, float x, float y ) {
 
 	if ( this->CheckSensor ( deviceID, sensorID, MOAISensor::TOUCH )) {
 		this->WriteEventHeader ( deviceID, sensorID, MOAISensor::TOUCH );
-		float time = USDeviceTime::GetTimeInSeconds();
+		float time = ( float )USDeviceTime::GetTimeInSeconds ();
 		MOAITouchSensor::WriteEvent ( this->mInput, touchID, down, x, y, time );
 	}
 }
